When there are multiple instances of the session id cookie present (because context paths in which the sessions exist have names that are proper prefixes of each other, like "/", "/foo", and "/foo/bar"), treat the *first* cookie value specified as the relevant one for this request, rather than the *last* one as presently done.
This is based on the RFC 2109 requirement that the browser is supposed to
send such cookies with the most specific path first, and this matches the
"longest match win" rule of how Tomcat selects which context will process
the request.

This patch was inspired by a Tomcat 3.2 patch proposed by Paul Frieden
<pfrieden@dChain.com>.  A modified version of his proposed patch that
provides equivalent functionality for Tomcat 3.2 will be committed next.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285333 13f79535-47bb-0310-9956-ffa450edef68