Migrate the handling of session id cookies from SessionInterceptor to StandardSessionInterceptor.
When there are multiple session ID cookies, take the *first* one as the
one that belongs to us, because of the rule in RFC 2109 that clients must
specify such cookies with the most specific "path" attribute first.  This
corresponds to the servlet spec rule that we take the longest prefix match
when selecting which context will handle a particular request.  (This is
slightly different than the original proposed behavior, and matches
behavior just added in Tomcat 4.0 as well.)

Submitted by: Paul Frieden <pfrieden@dChain.com>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282637 13f79535-47bb-0310-9956-ffa450edef68