(Porting a change from the 3.3 tree)
If Tomcat is run under a security manager, and a remote client tries to
connect that is not allowed by the security policy in place, log and
ignore the java.security.AccessControlException rather than shutting down
the endpoint.  This is done in a manner that should still be JDK 1.1
compliant.

Submitted by:	Glenn Nielsen <glenn@voyager.apg.more.net>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282666 13f79535-47bb-0310-9956-ffa450edef68