Improve backwards compatibility of servlet 2.2 applications running under Catalina, by conditionally accepting context-relative pathnames that do not start with "/".  To enable this acceptance, you must explicitly reference the public identifier of the 2.2 DTD in your web.xml file:
  <!DOCTYPE web-app
    PUBLIC "-//Sun Microsystems, Inc.//DTD Web Application 2.2//EN"
    "http://java.sun.com/j2ee/dtds/web-app_2_2.dtd">

(in other words, the standard DOCTYPE declaration for a 2.2 webapp).  The
paths without leading slashes are accepted on the following web.xml
elements:

* <jsp-file> subelement of the <servlet> element.

* <location> subelement of the <error-page> element.

* <form-login-page> and <form-error-page> subelements of the
  <form-login-config> element.

To encourage attention to this issue, a warning message will be logged to
the Logger associated with this Context, if such a value is accepted.

If you specify the 2.3 version of the DTD, attempts to specify such a path
without a leading "/" will be rejected.  This is based on the tightened
requirements that are specifically defined in the 2.3 specification.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285368 13f79535-47bb-0310-9956-ffa450edef68