A JSP page uses the JSP 1.2 XML syntax only if the jsp:root element is present.
-----
As reported on tomcat-dev:

Danno Ferrin:
>
> I am sorry to start this thread here, but I believe that jasper-4.0's
> behavior is in error.  The behavior I think the spec calls for in
> determining if a page is a JSP Document (xml jsp) or an XMl document
> with JSP markup is the presence or absence of a jsp:root element.
> [...]

Hans Bergsten:
I agree. JSP 1.1 says that everything that's not a JSP element is
treated as template text, and that any type of markup (text) language
can be used as template text. This means that a pure XML document,
with an <?xml version="1.0?> element at the top, is a valid JSP
page using the JSP 1.1 syntax.

The best (only?) way to identify a page that uses the JSP 1.2 XML syntax
is what you suggest: look for a jsp:root element. This should be
clarified in the JSP 1.2 spec.


PR:
Obtained from:
Submitted by: Danno Ferrin <shemnon@earthlink.net>
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285394 13f79535-47bb-0310-9956-ffa450edef68