Added a connector ( based on refactoring of DefaultCMSetter, etc) to set and manage the workdir.
New options to use a private directory inside WEB-INF ( according to the spec
a container is allowed to use it's own internal representation of a web
application ) - that would simplify the configuration of security
options and ( if the workdir is used to store persistent data like
compiled jsp files ) simplify the migration of webapplications
in a pool of containers.

This also allow people to completely replace the current behavior more easily,
by keeping all that's related with workdir in a single module.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282739 13f79535-47bb-0310-9956-ffa450edef68