Small change in util.log - expose the LogEntry.
I'm going to do another set of small changes to clean up the way loggers
are set up and maintained in ContextManager.

Among the changes, I would like to split Logger in LogManager and
LogHandler, use the package name to identify loggers. Probably
I'll try few performance enhancements ( recycle and pool LogEntry,
re-add the optional XML output format, try to make the interfaces similar
to the javax.logging so we can easily switch when the spec will be available)

I'll also try to port back the AccessLogValve.

( all this is low-priority, finishing the core is the most important thing)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282747 13f79535-47bb-0310-9956-ffa450edef68