First refactoring of MessageBytes.
The byte[] manipulation moved to ByteChunk, char[] to CharChunk. Both act
as "cursors" in an existing byte/char[] ( they can't resize and don't own
the buffer ).

The int/date conversion is cached, but it's just a cached value, it's no longer
"first class". Probably it's better to do the caching and conversion at a
higher level anyway.

In Headers, removed the Enumerations - as it was a bad choice and confusing.
It is still possible to enumerate a MultiMap, for use by the upper layer
( facade ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282749 13f79535-47bb-0310-9956-ffa450edef68