First round of refactoring for Handler/ServletWrapper. Reorganized the code to make sure the right order of calls in the right state is enforced.
It seems that Handler is still too complex, and a lot of servlet-specific code
( like handling of init, UnavailableException, etc) need to be moved in
the right place ( ServletWrapper). The problem is that ServletWrapper is too
complex.

I'll start cleaning ServletWrapper and then move init/destroy - Handler will
then be a very simple and maintainable class. I need your review on this one,
please send comments if you find any problem.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282762 13f79535-47bb-0310-9956-ffa450edef68