Stage 2 of Handler/ServletWrapper refactoring.
Split ServletWrapper in 2 smaller components - ServletInfo containing
all the information that web.xml is setting on a servlet, and
ServletHandler that deals only with the execution, init, destroy of the
servlet.

Moved LoadOnStartup into facade, since it's specific to internal details
of the servlet/jsp implementation ( the JSP init needs more work ). This
prepares for more simplifications in Handler.

Removed unused fields in Handler - load on startup, path, etc - all are
part of the ServletInfo, and not used by core.

Added comments on JspInterceptor - it's on the list for documentation/refactoring.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282768 13f79535-47bb-0310-9956-ffa450edef68