Another step in Handler/ServletWrapper refactoring - moved init, destroy and all related code to ServletHandler.
The big advantage is that now Handler is very simple, it deals only with
service() ( and we go back to the Apache/IIS/NES model where the handler
is a simple object ). All lifecycle ( init, etc) is present in the interceptor,
and duplicating it in Handler was confusing and resulted in many errors.

The Handler is a part of the extension mechanism of tomcat - a module
( Interceptor ) can define a number of handlers ( like StaticHandler,
ServletHandler) to generate the output ( either directly or via another
mechanism - like servlets, jsps )

I think this is the right direction, please review it and send any comments.
( there are few more simplifications, but this commit has only the code
move and associated fixes )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282774 13f79535-47bb-0310-9956-ffa450edef68