One more step in Handler refactoring. We are very close, I think Handler starts to be simple enough and ServletWrapper is now at least readable ( that was one of the most complex parts of tomcat3).
In the process I also fixed a number of bugs ( it's amazing what you
find when the code is simpler ) and added few improvements on the
JspInterceptor ( now the main target for refactoring )

- added get/setModule - the handler is associated with an interceptor the
same way as in apache, etc. The interceptor creates the handler and sets
it ( properties, debug ) - Servlet22Interceptor and JspInterceptor are doing
this for servlets/jsps

- many fixes in JspInterceptor - the error handling had few problems,
same for reloading.

- changed JspInterceptor to use the same mechanism for reloading with
servlets - now expire checks are done periodically, not for each
request.

- started to simplify and better organize JspInterceptor

- fixed bugs related with jsps declared in web.xml

- added a small test for jsp declarations

- use a prefix for the names declared by JspInterceptor, to avoid
conflicts with web.xml names

- changed the Context logger to report the context path

- removed accounting from Handler - a better mechanism is needed, and
it's not a priority.

- that simplifies a lot the Handler, and make it possible and easy to
pass options using server.xml

- removed "origin" from handler - it's implied by Module

- removed Context from handler - it's used only in ServletHandler.

- Small hack in ServletInfo so that servlets declared by WebXmlReader will
have Servlet22Interceptor as "owner".


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282777 13f79535-47bb-0310-9956-ffa450edef68