Fix a buffering issue that was causing one Watchdog-4.0 test failure (SetBufferSize_1Test), and also could cause grief for applications as well.
The issue:  Several methods on ServletResponse are required to throw an
IllegalStateException if any output has been written to the output stream
or writer, even if the response has not been committed to the client yet.
The original implementation simply returned a PrintWriter that wrapped an
OutputStreamWriter -- to do the character encoding conversions -- around
the ServletOutputStream implementation that writes directly to the
response buffer.

The problem:  The java.io.OutputStreamWriter implementation includes a
non-configurable buffer (8kb per instance in the Sun JDK!!!).  When you
"print" output to this buffer, the ServletResponse does not know it has
happened, and you get incorrect results from calls like setBufferSize().

The solution:  Create a customized PrintWriter, to be returned when you
call response.getWriter().  This version causes the OutputStreamWriter to
be "flushed" to the underlying ServletOutputStream, which simply writes
directly to the response buffer, on every write() or print().  Now, the
response is aware of what bytes have *really* been written.  In addition,
response.reset() and response.resetBuffer() can now work correctly even
when using a writer.

The performance impact of this change is certainly worth being concerned
about, but correct functionality has to come first -- now we can optimize.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285452 13f79535-47bb-0310-9956-ffa450edef68