updated to latest ant support. these are the same files from the head of cvs.
one MAJOR complaint i have is that you can't have a copy of the head of
cvs built and and another copy of the 2.3 tree because they don't build into
their own directories. that should be fixed, but i'm afraid it will
break a bunch of other dependent code in tomcat if i fix that. sigh. someone
(ie: craig) with more expertise on this should probably fix it. :-)

-jon


git-svn-id: https://svn.apache.org/repos/asf/tomcat@265915 13f79535-47bb-0310-9956-ffa450edef68