More fixes for exception handling.
It now works correctly when an included serlvet catches an exception and
either handles it or rethrows a different exception.

It also works correctly when an included servlet throws an permanent
UnavailableException.  If "includers" do not handle the exception, they are
not made permanently unavailable too!

The idea of saving the original exception on Response and not overwriting it
doesn't work when exceptions are caught and handled or a different
exception is thrown. As a result, Handler.saveError() now always saves new
exceptions on the Response, not updating the information if the exception
is the same as the one already set.

Removing the errorException property from Request since it is no longer
needed.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282791 13f79535-47bb-0310-9956-ffa450edef68