Add the initial version of a unit testing suite for Tomcat 4.0.  The test client is a custom Ant task modelled after the GTest task used by the Watchdog tests, but which uses HttpURLConnection instead of a low level socket connection for processing.
The intent of this suite is to host unit tests for Tomcat-specific
functionality, rather than the spec compliance that is checked by
Watchdog.  As such, the tests rely on specifics of the way that a default
Tomcat 4.0 installation is laid out.

Currently, the tester suite is *not* integrated in the top-level build
scripts for Tomcat 4.0.  To build and deploy it to the "build" version,
execute the following commands from the top level directory of the source:

	cd tester
	./build.sh deploy-main

and then restart Tomcat 4.0.  To execute the tests, execute:

	$CATALINA_HOME/bin/tester.sh


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285495 13f79535-47bb-0310-9956-ffa450edef68