More changes for Apache2.0.
Now it works for pages<4k, there is still a problem with ap_rwrite()
when we send larger pages ( probably we need to call a different apache
function )


Note that APXS changed and now uses libtool in a strange mode ( i.e. apxs *.c
will create a module for each c file instead of the previous behavior when
it creates a module containing all c files ).

That means you have to use Makefile.linux to build - I hope Dan or Henri will
take a look and write a "real" makefile.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282799 13f79535-47bb-0310-9956-ffa450edef68