Check for null value before invoking method.
From email sent by Brian Bucknam:

It's a long story, but I'm working on a project where Jasper 3.x is embedded
inside a servlet, which can then be deployed to the container of our
customer's choice.  The servlet uses JSP files as templates, which is where
Jasper comes in.

In this type of environment, sometimes thing can go really wrong, and the
compiled JSP might, in some cases, fail to get a JspFactory, PageContext, or
JspWriter.

If any of _jspxFactory, pageContext, or out fail to be created, the catch{}
and finally{} clauses just throw NPE's.

PR:
Obtained from:
Submitted by:  "Bucknam, Brian" <Brian.Bucknam@webgain.com>
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282802 13f79535-47bb-0310-9956-ffa450edef68