Bug fix: bug report #269 -- java.io.UnsupportedEncodingException when processing JSP
From the bug report:

"A "java.io.UnsupportedEncodingException: UTF8" is thrown when generating
the servlet for a JSP file when the Kaffe VM is used. The "UTF8" encoding
name is hardcoded in the Compiler class as the encoding for the generated
servlet source code file. I believe the reason for using "UTF8" as opposed
to "UTF-8" (note the dash) is that this is the only name supported in
JDK 1.1. I suggest adding an init parameter to JspServlet for setting
the encoding name when "UTF8" doesn't work."

Given that there are several possible representations of Unicode data
(UTF-8, UTF-16, UTF-32), and given that the supported encodings vary
between different implementations of the Java platform,
the best way to tackle this is probably as suggested by Hans.


PR:
Obtained from:
Submitted by:	Hans Bergsten ( hans@gefionsoftware.com )
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282803 13f79535-47bb-0310-9956-ffa450edef68