Extend the unit testing framework with wrappers for the request and response objects that can be configurably added with a Filter defined in the web.xml file.  Now, where it makes sense, we can run the wrapped and unwrapped version of the same test, to ensure that introducing wrappers does not cause any functional problems.
In addition, a logger that uses static methods has been added to allow
messages from various components of a test (such as filters and wrappers)
to be accumulated and added to the response message.

TODO:  TesterHttpServletRequestWrapper needs to be written, and
WrapperFilter needs to use it if so configured.  This will allow the
wrapped version of the GetQueryString01 test to be uncommented.

TODO:  Golden file support needs to be added to the test client, so that
the entire response can be validated (not just the first line, as is done
on "outContent".  This will be needed to validate information added to the
response from the static logger.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285509 13f79535-47bb-0310-9956-ffa450edef68