Add a backwards compatibility hack to the way that web application deployment descriptors are parsed.
Tomcat 3.x accepted <url-pattern> arguments (inside a <servlet-mapping> or
a <security-constraint>) that did not conform to the syntax requirements
of the Servlet 2.2 specification, because it would explicitly add a
leading slash if it was omitted.  Now, Tomcat 4.0 will accept such a
pattern, with a warning message to the log, *only* on applications that
declare themselves to use the Servlet 2.2 DTD.

For a 2.3 web app, Tomcat 4.0 enforces the syntax rules on <url-pattern>
arguments, which are spelled out explicitly in the spec.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285511 13f79535-47bb-0310-9956-ffa450edef68