Fixed ErrorHandler - showDebugInfo was removed from ContextManager as an "explicit" property. Larry - we can add it back, I just wanted to test the setProperty mechanism.
The fix also allow contexts to override the setting in context manager -
you can set it now at context level.

What's interesting is that it does that without any special code in core -
any new attributes can be added to <context > and <contextManager> tags.

We should add properties to Context and ContextManager only if they are
very generic ( like "home" ).

Another method to do the same thing is to use per/context interceptors,
or settings on the interceptor ( for example showDebugInfo="true" on the
error handler ). This is better if the attribute is specific to a particular
module ( like session saving in a file ). For showDebugInfo, it's likely that
other modules will use it.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282829 13f79535-47bb-0310-9956-ffa450edef68