Transformed the "automatic server config" generators in modules.
Those classes were an attempt to simplify the server configuration for
"integrated" mode. The problem is that tomcat can add/remove applications
at runtime, modules like AutoSetup can also add applications.
Another problem is that users still need to edit the files.

One thing is clear - the code in doesn't work as it should.

By turning the generators into modules we allow access to all server hooks -
it will be possible to regenerate the configs ( and maybe restart the server)
when a context is added/removed, even at runtime.
We'll also have more flexibility in configuring ( we can add only the
generator we need, and pass the location of apache/iis/nes, etc ).

The code is low-priority - if ajp1x-based autoconfiguration will work this will no longer be needed. I think there is a lot of work to be done to
fully integrate with the server ( like using the native-authentication, etc),
and I don't think that can be done in 3.3 timeframe - but later, as a module.

I would even propose removing them completely from the 3.3 distribution,
since they are not essential ( and were never completed )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282844 13f79535-47bb-0310-9956-ffa450edef68