Minor fixes ( including the build :-), preparing to finish RequestUtil ( the last "intermediary" class - neither core nor util nor module ).
After that is done, tomcat 3.3 will consist of only 3 components:

- the "core" - 9 classes ( including TomcatException ): basic representation
for request, response, location, application, server, module.

- modules ( tomcat/context and tomcat/request modules will be moved to
tomcat/modules and organized by type, again as in Apache 2.0 ). ( about 30
modules so far - probably 20 are important )

- utilities ( similar with APR - more work is needed to optimize and improve
the documentation, but that's a different story, not tomcat3.3 )

( the servlet2.2 is a particular module/group of modules ).

( One more step would be to remove/move out all non-essential modules
( i.e. "features" that are not required for normal operation ) - but I don't
like removing existing features, probably we can leave them in.)

Remaining items for this year: Parameters, Sessions, maybe Encoding (
making it a module - the rest can be done independently ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282863 13f79535-47bb-0310-9956-ffa450edef68