Finish ( I hope ) refactoring of session management.
- ServerSession moved to core. It's no longer final - session manager modules
can extend and replace the functionality.

- added few more properties ( for symetry with other classes in core):
notes, context, state, facade.

- Replaced the "newSessionRequest" hook with 2 other hooks: findSession
and sessionState. This allow modules to control most aspects of session
processing.

- Servlet22Interceptor now corectly handles servlet-specific aspects of
session management - binding events, etc. That happens for both
expired sessions and reloading.

- Small improvement in reloading: we'll set a context note with the old
class loader, and on session reloading we replace only the objects loaded
with the old loader. That means an object that is loaded with the main
class loader ( like DatabaseConnection ) will be valid after reload.
( should this be optional ? )

- SessionId will now verify if the session is valid for both cookies and
URL rewriting ( used to be only for cookies ). It'll also call "touch" -
SimpleSessionManager will deal only with the 2 session hooks, no more
intervention in the mapping hooks.

- The ServerManager is now part of SimpleSessionStore. A session
plugin is formed from a session manager and a tomcat 3.3 interceptor.

Another small change: if the module implements "registerHooks", it'll
be allowed to register the hooks itself, instead of using the automated
code in Hooks. That means it can insert itself in front, end or after
another module ( apache 2.0 style ) - by default the modules are inserted
automatically at the end.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282867 13f79535-47bb-0310-9956-ffa450edef68