Renamed Handler.doService() to Handler.invoke().
Added comments and get/set Next/Prev for handlers that act as Valves.
Same thing can be implemented in a ValveHandler ( that will be part
of the module using valves ), but it's better to make clear in the
core Handler that this mechanism is possible ( and simplify the life of
whoever wants to use/reuse valves ). It's also a safety feature for tomcat
extensibility - in case something is missing in the current hooks ( and
we don't want/can't add more hooks )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282872 13f79535-47bb-0310-9956-ffa450edef68