- Added showDebugInfo property to the ErrorHandler module - it'll act as default ( explicit properties are easier to document, and it's better to set the properties on the module if it's specific to the module )
- Bug fix: the buffer will be reset before displaying the exception, even for
the default handler. It shouldn't be any distinction between the "default" error
handler and a user-defined handler, and this helps if the buffer may have
garbage ( open tags, etc ).

- Bug fix: reset the exception state before calling the error servlet. This prevent
the display of "error in the error handler"

- Bug fix: if( ! (ex instanceof ... ) -> if( ex!=null && ! ( ex instanceof ... )

- Bug fix: add the <head>, <body>, etc only if the buffere is "clean". The previous
test ( getBytesWritten() ) was bad, the correct test is "isNew()" ( because both
chars and bytes can be written to the buffer ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282891 13f79535-47bb-0310-9956-ffa450edef68