Merged the jasper changes from the release branch (3.2.1+).
This is an automated merge ( cvs update -j ), with just a bit of manual conflict
resolutions. From what I can tell it worked pretty well - and probably this is the
best way to go to keep jasper in sync.

For tomcat we still need to check the remaining commits that were not yet merged into
the main branch - this can't be automated because of the code reorganization and
refactoring.

For people making changes into 3.2.1 -> PLEASE  TRY TO MERGE THE FIX INTO THE MAIN
BRANCH TOO !



PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282895 13f79535-47bb-0310-9956-ffa450edef68