Few minor changes/fixes ( done long ago ) to better guess tomcat.home ( useful if you embed tomcat or use "java -jar" style of startup ). ( everything is fine if TOMCAT.HOME is set or PWD is TOMCAT.HOME, but this is not allways the case - that's what we fix here. )
IntrospectionUtil looks in the classpath ( assuming jars are not moved
around too much ) and finds the location of tomcat_core.jar, and sets
tomcat home based on that. It's the equivalent of the .sh script.
This is needed on Windows if executable jars are supported. There is
more testing to be done.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282934 13f79535-47bb-0310-9956-ffa450edef68