Varius cosmetic changes to the testing component of the admin webapp.
- Option to do "external" formating ( i.e. the test results are stored,
and the JSP page can do whatever formating it wants ). Old style is
still supported ( text, simple html generated by GTest )

- A simple tag to set the defaults for HttpClient and GTest ( avoid quite
a few keystrokes - host="${host}", etc for each test )

- test.jsp will display the FAILED tests first ( easier to read )

- the test now detects HTTP/0.9 errors ( 9 tests are now failing, mostly
because of a bug in HTTP/0.9 support - not hard to fix, I think the bug
is also present in 3.2 ). The old GTest didn't implemented 0.9 support
corectly, and accepted 1.0 responses for 0.9 requests. Few of the tests are
wrong from the same reason ( expect status on 0.9 requests )

- fixes, more options, etc


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282939 13f79535-47bb-0310-9956-ffa450edef68