Begin updating for next Tomcat.
Added items for some of the changes from Tomcat 3.2 that have been
completed.  These try to summarize some of the information found in the
changes3.3 file.  These additions are identified by the release being specified
as 3.3.

Append a section for adding action items related to the release.  I'll try to
flesh this out soon.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282940 13f79535-47bb-0310-9956-ffa450edef68