Added code and build changes to allow us to run watchdog using a simple web page.
The "right" way to run watchdog is by using the provided scripts. The current
mechanism ( built using "ant watchdog-web-based" and run by accessing
http://localhost:8080/admin/test/watchdog-jsp.jsp , watchgod-servlet.jsp )
is intended to increase the test pool by making it easier for people with
special platforms to run tests. ( example - I don't think a Mac will run
either .sh or .bat files ).

Running watchdog as web application is mostly a hack - for example the
servlet tests are based on moo ( remember moo ? ), and the trick used is
an <ant> java task that runs it ( with fork=true ), so that the output
goes to the ant logger ( AntServletLogger in this case ). This is the
trickiest part, for jsps GTest is used ( and we use <replace> to patch
 in the new GTest ).

This is not production-quality or even close - it's just a quick way to
get things started ( testing and adding regression tests to the /test
application ). The nightly builds will include both /test and watchdog,
in order to allow people to check tomcat on their platform. Running
the scripts is still the "correct" way to run watchdog.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282943 13f79535-47bb-0310-9956-ffa450edef68