2 fixes in the Http implementation: - do case-sensitive compare for the http method ( GET works, get doesn't ) - right HTTP 0.9 support
Also, a fix in the matcher ( the responseCode is not matched, needed to
deal with 0.9 requests ).

I also commented out or changed a number of tests ( POST doesn't exist
in 0.9, so need to add the protocol ).

Right now all but one tests are passing.
The failing test works fine with JspInterceptor ( if useJspServlet="false"
is used ), but fails with JspServlet.

( the current default is to use JspServlet, in order to test and make sure
it is supported and works fine ).

The failing test is test/jsp/classTest.jsp, with the error:
java.lang.IllegalAccessError: try to access class jsp/_0002fjsp_0002fclassTest_0002ejspclassTest_jsp_0$JSPTest$Inner from class jsp/_0002fjsp_0002fclassTest_0002ejspclassTest_jsp_0


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282947 13f79535-47bb-0310-9956-ffa450edef68