If an expanded context is already defined, try to reload the context instead of recreate it.  This has a better chance of preserving the context settings. :-)
Note: Context reloading is not currently successful at completely avoiding
the bug.  If you try to execute the admin/test/test.jsp when the admin.war
wasn't already expaned, the request will result in a 500 error. The
ServletContext for the JspServlet is still carrying a reference to the
old context and its old classloader.  This old classloader doesn't have
WEB-INF/classes in its urls[] array. If this problem occurs, restart Tomcat.
Since the admin.war is already expanded, the problem is avoided.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282949 13f79535-47bb-0310-9956-ffa450edef68