Replaced <Logger> with the LogSetter, now everything works the same way without any special hacks.
The comments on logging moved to the LogSetter - I'll also migrate
more of the comments from server.xml to the modules.
The idea is that module documentation ( that should be located in
modules themself - or in the user documentation ) should be part of
the docs, not in xml comments. server.xml might get re-generated by
tools and the comments are hard to preserve.

We can add an explicit description="..." to each module, to explain
what it's doing, but I don't think server.xml is the right user documentation

( if we decide to use it as user doc, we should add comments for all
modules )

( also fixed the class name in LogSetter - bad commit )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283005 13f79535-47bb-0310-9956-ffa450edef68