Fix the System.getProperty("line.separator") problem in JspWriter. ( it was causing most of the failures in the nightly tests for sandboxed tomcat ).
This adds another dependency between jasper and tomcat.util ( the first is
tomcat.util.log, now tomcat.util.compat ). Since both packages are completely
standalone and can be bundled with jasper, I don't think it'll be a problem.

( of course, a better solution would be to have the general-purpose tools
in a org.apache.util.{log,compat}, there is no reason to keep them in tomcat )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283138 13f79535-47bb-0310-9956-ffa450edef68