Modify the tester suite execution scripts to reflect the change in JAR file locations for the XML parser that is required.
Add unit tests to validate correct processing of <error-page> elements in
the web application deployment descriptor, and correct handling of
forwarding to error pages on a particular HTTP status code, as well as
when a particular exception is thrown.  This set of tests was created to
validate the behavior of error handling, in response to Bugzilla Bug#644.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285663 13f79535-47bb-0310-9956-ffa450edef68