Another change related with the profiles/lib reorg - remove the dependency between core and DependManager.
As with the previous changes, DependManager is just an implementation
choice made by the Reload module ( and few other modules adding dependencies).
There are other ways to implement reloading ( modified class loaders, etc ),
and there is no reason to "force" one way in the core APIs.

DependManager is specific to a module ( or few ), but not required
for normal tomcat operation ( other choices are possible as well ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283213 13f79535-47bb-0310-9956-ffa450edef68