Initial version of ProfileLoader.
It's a very simple class - it will read a profile description ( the
config attribute ) and store the set of modules and their config.

When a context is loaded with a propery "profile" matching the profile
name, all the modules will be added to the context.

For example:
 <ProfileLoader config="conf/profile.xml" name="foo" />

 ...

 <Context profile="foo" ... />

The profile has 2 roles:
- simplified config ( you don't have to specify Jdbc auth + specific logger
options for each module )

- ( later ) allow groups of contexts to share some classes ( i.e. form a
hierarchy of class loaders ) - this is very important for inter-context
communication, without requiring libs shared by few apps to be
forced in the CLASSPATH ( at global level ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283216 13f79535-47bb-0310-9956-ffa450edef68