Added the top-level directories for the contributed modules and apps that will not be included in the release ( either because they are not ready or as a choice to keep the release focused).
This will allow us to check in the new adminUI and the digested auth.

Modules here should be considered "revolution"-like, and are not part
of the release ( or optional/ in ant terms ).

The main goals are to avoid featurism in the released tomcat, while
allowing people to add the functionality they need ( without impacting
the release cycle )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283217 13f79535-47bb-0310-9956-ffa450edef68