Another code move for simpler dependencies.
This time with a more significant change - the aaa modules must set
the Principal ( instead of relying on core to create one ).

The Principal is carying more information than the String user, and
in general we should use Principal whenever possible. ( for example
by checking the class name of the principal you can find who did
the authentication  )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283220 13f79535-47bb-0310-9956-ffa450edef68