Port fix from tomcat_32 by Marc Saegesser.
Strip any double-quote characters from around the charset value
in the Content-type header.

Coincidentally, there was a recent thread on the HTTP WG mailing
list about this topic and the conclusion was that that specification
does support enclosing the value of the charset attribute in quotes
(although I'm at a loss as to why anyone would want to).

Reported by Rick Rupp (rick.rupp@merant.com).

PR: 650/ Buzzilla 364


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283353 13f79535-47bb-0310-9956-ffa450edef68