For Tomcat 4.0, fix the security vulnerability reported by Hiromitsu Takagi.
The problem actually has nothing to do with JSP pages per se -- it is due
to the fact that the original request URI is included in the response on
many of the standard error pages produced by Tomcat.  In the case at hand,
it is the standard message for error 404 (not found).

You can prove that it is not related to JSP by trying *any* URI that
includes JavaScript code, and triggers a 404, such as:

http://localhost:8080/examples/<SCRIPT>alert(document.cookie)</SCRIPT>.xyz

The fix is to filter the message string included in the response, so that
characters sensitive to HTML are rendered as their corresponding escape
sequences (such as translating "<" to "&lt;") so that the browser will
render them rather than execute them.

WARNING:  Web application error pages that naively display the request URI
in their output can be subject to this same kind of problem.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285724 13f79535-47bb-0310-9956-ffa450edef68