Disable creation of session facades. A better solution is implemented in 3.3, where the same facade is returned during servlet execution.
The session facade was added for security - to prevent the servlet from
accessing public methods and to prevent the servlet to "keep" the reference
and use it later ( when the object may be associated with another session ).

Securing tomcat and support for untrusted apps is a long process, and 3.2
has other places where untrusted apps could hold references to recycled
objects - as long as the HttpServletRequest is recycled there is no
point in beeing paranoic about the session.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283377 13f79535-47bb-0310-9956-ffa450edef68