Any significant error that is encountered while starting a Context initially, or as part of a reload() call, will now cause the corresponding web application to remain unavailable, rather than allowing it to operate in a partially initialized state.
The definition of "significant error" includes failure in initializaing any
Listener or Filter that is defined in the web.xml file.  Whether or not such
errors should cause the application startup to fail or not is currently under
discussion in the JSR-053 expert group.

PR: Bugzilla #1129 (status changed to RESOLVED/REMIND)
Submitted by: Kevin Jones <kevinj@develop.com>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285792 13f79535-47bb-0310-9956-ffa450edef68