One more time to address the double decode security hole.
Trying to detect the URLFileConnection bug turned out to be problematic so I've decided for now to just disallow all resource names with URL escapes
in them.  This certainly prevents the security hole but does preclude
some resources that with valid escapes in them.  This will be noted as
a Tomcat limitation.

Thanks to Costin for the simplifying idea.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283436 13f79535-47bb-0310-9956-ffa450edef68