Initial version of a Realm implementation for Tomcat 4.0 that utilizes a directory server (accessed via JNDI) to perform user authentication and access control for container-managed security.  This code is based in large part on the proposed code (on TOMCAT-DEV) by John Holman, with some additions and refactoring by me.
Use of this realm would be configured in "conf/server.xml" by an entry
like this to connect to an LDAP server on the same host that Tomcat is
running on:

  <Realm className="org.apache.catalina.realm.JNDIRealm"
    connectionName="admin-username"
connectionPassword="admin-password"
     connectionURL="ldap://localhost:389"
       userPattern="cn={0},dc=mycompany,dc=com"
      userPassword="userPassword"
          roleBase="dc=groups,dc=mycompany,dc=com"
          roleName="cn"
        roleSearch="(|(uniqueMember={0})(member={0}))"
       roleSubtree="false"
  />


TODO:  Update the configuration documentation to describe the above.

TODO:  Support an operational mode where the Realm attempts to bind to the
directory server using the user's username and password (instead of a
system administrator username and password).  This is a different enough
style that it probably should be a separate implementation class.

TODO:  Support connection pooling (for both this and JDBCRealm) so that
the authenticate() method does not have to be synchronized.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285868 13f79535-47bb-0310-9956-ffa450edef68