Restore the ability to save and reload active sessions across a web app restart.  This was broken by the refactoring of the load() and unload() calls that was recently done in the session manager code.
The problem was a ClassNotFoundException (on a class found in WEB-INF/classes)
when the session manager attempted to load the saved session.  This was in turn
due to the fact that the URLStreamHandler for the directory context was not
initialized when this attempt was happening.  Rearranging the order of the
steps fixes the problem.

PLEASE run the entire "tester" suite, as well as the Watchdog 4.0 tests, to
ensure that we do not introduce regressions like this on future changes.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285889 13f79535-47bb-0310-9956-ffa450edef68