For example, in the jakarta-taglibs xsl tag:
<xsl:apply xml="<%= xmlFile %>" xsl="<%= xslFile %>"/>

A bug resulted when the first element value "<%= xmlFile %>" matched up with the first attribute's properties. In this case, it would have been "nameXml" as defined in the xsl.tld. "nameXml" has its rtexprvalue property set to false.

The bug fix makes sure that attribute properties match up with the correct element.

Bugzilla: #811


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285895 13f79535-47bb-0310-9956-ffa450edef68