Fix server.xml with the right name of the attribute.
Delay Random creation ( each random creates a thread )

That concludes the fix of 1418, no hanging threads after shutdown.

Note that shutdown/init is not completely debugged or tested - but at least there
are no gross errors and kind-of works. In time things will improve ( the assumption
is you don't do a soft restart every hour ).

Soft restart is better in many cases - like when you embed tomcat in an application,
as it doesn't require to restart the java vm and should clean up all the resources
and return to a stable state.

Of course, that requires that all modules respect the server state and clean
up resources.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283462 13f79535-47bb-0310-9956-ffa450edef68