[PFD2-9.9.2] - Implement the revised algorithm for looking up the error page associated with an exception thrown by the top-level servlet (first match wins): - Look for an error page with this exact exception class - Look for an error page for progressive superclasses of the exception class - If this exception is a ServletException with a rootCause property,   rerun the above two steps on the rootCause exception - Display the container's standard error page.
NOTE:  Currently, four of the ErrorPage tests in tester will fail, because
they assume that the exception forwarded to the error page will be the root
cause exception, rather that the ServletException.  I've submitted a request
for clarification to know which behavior is correct, and will change either
the logic or the tests accordingly.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285936 13f79535-47bb-0310-9956-ffa450edef68